LPF-190 - use BeforeAll to avoid spinning up service for every individual test
