+++ .github/workflows/build-and-test.yml
@@ -29,8 +29,6 @@ jobs:
         run: |
           cd sourceCode
           mvn -B -DskipTests clean package
-          ls -lart
-          pwd
 
       - name: Add source code jar to local maven repository
         run: mvn install:install-file -Dfile=sourceCode/target/pay-for-legal-aid-0.0.1-SNAPSHOT.jar -DgroupId=uk.gov.laa -DartifactId=pay-for-legal-aid -Dversion=0.0.1-SNAPSHOT -Dpackaging=jar
