+++ src/test/java/uk/gov/laa/pfla/RunCucumberTest.java
@@ -1,9 +1,14 @@
 package uk.gov.laa.pfla;
 
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.TestInstance;
 import org.junit.platform.suite.api.ConfigurationParameter;
 import org.junit.platform.suite.api.IncludeEngines;
 import org.junit.platform.suite.api.SelectPackages;
 import org.junit.platform.suite.api.Suite;
+
+import java.io.IOException;
 
 import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PROPERTY_NAME;
 
@@ -11,5 +16,24 @@ import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PROPERTY_NAME;
 @IncludeEngines("cucumber")
 @SelectPackages("uk")
 @ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "pretty")
+//The below allows us to just spin up the service once and not for every test
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class RunCucumberTest {
+    private ServiceManager serviceManager;
+
+    // This is a Junit BeforeAll.
+    // Note there is a Cucumber BeforeAll we can also use simultaneously if there is Cucumber Specific Setup needed
+    @BeforeAll
+    public void setup() throws IOException {
+        serviceManager = new ServiceManager();
+        serviceManager.startService();
+    }
+
+    @AfterAll
+    public void teardown() {
+
+        if (serviceManager != null) {
+            serviceManager.stopService();
+        }
+    }
 }
