+++ src/test/resources/uk/gov/laa/gpfd/health_api.feature
@@ -1,6 +1,6 @@
-Feature: Call Health Endpoint
+Feature: Call Actuator Endpoint
 
-  Scenario: Return 401 response
+  Scenario: Return 200 response
     Given the service is running and we are not logged in
     When it calls the actuator endpoint
-    Then it should return a 401 response
+    Then it should return a 200 response
